Khushiyaan-Foundation-Dashboard-documentation/requirements.txt
dash
pandas
plotly
numpy
gunicorn==23.0.0
dash-bootstrap-components===2.0.3
openpyxl==3.1.5
python-dotenv
flask-login
gspread
oauth2client
Pillow
reportlab
PyPDF2